RISC-V: Add experimental support for Zvabd extension
This patch adds experimental support for the RISC-V Zvabd (Vector Absolute
Difference) extension, based on version 0.5 of the proposal. The
implementation includes instruction patterns and test coverage for
auto-vectorization.

The Zvabd instruction set specification can be found in the fast-track proposal:
https://lf-riscv.atlassian.net/wiki/spaces/VXXX/pages/166690866/Fast-Track+Proposal+for+integer+vector+absolute+difference+instructions

Note: This is early-access support for evaluation and may change as the
Zvabd proposal evolves.

gcc/ChangeLog:
    * config/riscv/autovec.md: Add auto-vectorization patterns for Zvabd instructions.
    * config/riscv/riscv-ext.def: Add Zvabd extension entry.
    * config/riscv/riscv-ext.opt: Add Zvabd option.
    * config/riscv/vector-iterators.md: Update for Zvabd.
    * config/riscv/vector.md: Add Zvabd patterns.

gcc/testsuite/ChangeLog:
    * gcc.target/riscv/rvv/autovec/zvabd-1.c: New tests for Zvabd.

Co-authored-by: Yuke Tang <tang.yuke@zte.com.cn>
Signed-off-by: Zhongyao Chen <chen.zhongyao@zte.com.cn>